% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{Character3Grams}
\alias{Character3Grams}
\title{Character 3-gram extraction.}
\usage{
Character3Grams(text)
}
\arguments{
\item{text}{The text.}
}
\value{
A document term matrix (sparse Matrix).
}
\description{
Computes the document term matrix of character 3-gram.
}
