% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\docType{data}
\name{features}
\alias{features}
\title{Features.}
\format{An object of class \code{module} (inherits from \code{list}) of length 19.}
\usage{
features
}
\description{
Module containing functions for individual simple text feature
extraction.
}
\details{
Most functions have a single \code{text} parameter. The module
contains the following functions:

\describe{
  \item{\code{Stopwords}}{Number of stopwords. Uses two optional
  parameters: \code{Tokenize} which is the word tokenizer to use
  and \code{stopwords} which is the list of stopwords to use.}
  \item{\code{Tokenize1}}{First tokenizer available for
  \code{Stopwords}.}
  \item{\code{Tokenize2}}{Second tokenizer available for
  \code{Stopwords}.}
  \item{\code{StopwordsRatio1}}{Ratio of stopwords using \code{Tokenize1}}
  \item{\code{StopwordsRatio2}}{Ratio of stopwords using \code{Tokenize2}}
  \item{\code{Caps}}{Number of uppercase letters.}
  \item{\code{CapsRatio}}{Ratio of uppercase letters.}
  \item{\code{SpecialChars}}{Number of special characters.}
  \item{\code{SpecialCharsRatio}}{Ratio of special characters.}
  \item{\code{Numbers}}{Number of digit characters.}
  \item{\code{NumbersRatio}}{Ratio of digit characters.}
  \item{\code{Words}}{Number of words.}
  \item{\code{AverageWordLength}}{Average word length.}
  \item{\code{LastCharCode}}{Boolean for the use of a code character at the
  end of the text.}
  \item{\code{LastCharNL}}{Boolean for the use of a natural language boolean
  at the end of the text.}
  \item{\code{First3Chars}}{Returns the first three non white characters.}
  \item{\code{First3CharsLetters}}{The number of three first non white
  characters that are letters.}
  \item{\code{Emoticons}}{The number of emoticons}
  \item{\code{StartWithAt}}{Boolean for the use of @ at the start of
  the text.}
}
}
\keyword{datasets}
