% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{NLoNPredict}
\alias{NLoNPredict}
\title{NLoN Model.}
\usage{
NLoNPredict(model, text, lambda = NULL, type = "class",
  features = TriGramsAndFeatures)
}
\arguments{
\item{model}{A glmnet model as returned by \code{\link{NLoNModel}}.}

\item{text}{A character vector containing the text to predict.}

\item{lambda}{Lambda parameter to pass to
\code{\link[glmnet]{predict.glmnet}}.}

\item{type}{Type of prediction made by
\code{\link[glmnet]{predict.glmnet}}.}

\item{features}{A function computing the feature values (a matrix,
list of numeric vectors or data.frame) or a list of functions
computing individual feature values.}
}
\value{
The output of \code{\link[glmnet]{predict.glmnet}}.
}
\description{
Train a NLoN model using glmnet.
}
\seealso{
\code{\link[glmnet]{predict.glmnet}}
}
