% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{TriGramsAndFeatures}
\alias{TriGramsAndFeatures}
\title{3-grams and feature extraction.}
\usage{
TriGramsAndFeatures(text)
}
\arguments{
\item{text}{The text.}
}
\value{
A sparse Matrix with text features and 3-gram.
}
\description{
Computes both 3-gram and simple text features.
}
\seealso{
\code{\link{Character3Grams}}

\code{\link{FeatureExtraction}}
}
